Move report endpoints to new module
