+++ backend/src/report/mod.rs
@@ -1,0 +1,8 @@
+//! Report API.
+//!
+//! Contains API endpoints to produce reports as PDF ('proces-verbaal')
+//! and XML using the EML_NL standard.
+
+mod api;
+
+pub use self::api::*;
